VM: modifs for most functions except : sti, ldi, lldi

+++ VM/src/ft_live_zjmp_fork_lfork.c
@@ -20,7 +20,7 @@
 
 int					ft_live(t_vm_mem *vm, t_ps *ps, int opcode)
 {
-	unsigned int	arg0;
+	int				arg0;
 	t_ps			*lst;
 	int				i;
 
@@ -32,7 +32,7 @@ int					ft_live(t_vm_mem *vm, t_ps *ps, int opcode)
 	lst = ps;
 	while (lst)
 	{
-		if (lst->uid == (int)arg0)
+		if (lst->uid == arg0)
 			break ;
 		lst = lst->next;
 	}
@@ -56,21 +56,17 @@ int					ft_zjmp(t_vm_mem *vm, t_ps *ps, int opcode)
 	i = -1;
 	arg0 = 0;
 	DEBUG ? ft_printf("launching ft_zjmp ...\n") : DEBUG;
-	//ft_printf("ps->pc 1: %d\n", ps->pc);
 	while (++i < DIR_SIZE(OP_TAB_INDEX(ZJMP)) && ((arg0 = arg0 << 8) || true))
 		arg0 += *(vm->mem + MEM_CIR_POS(ps->pc + OPCODE_SIZE + i));
-	//ft_printf("arg0 : %d\n", arg0);
-	//ft_printf("arg0 hex : %#x\n", arg0);
 	if (ps->carry == CARRY_TRUE)
 		ps->pc = MEM_CIR_POS(ps->pc + arg0 % IDX_MOD);
-	//ft_printf("ps->pc 2: %d\n", ps->pc);
 	return (EXIT_SUCCESS);
 }
 
 int					ft_fork(t_vm_mem *vm, t_ps *ps, int opcode)
 {
 	t_ps			*new;
-	unsigned int	arg0;
+	int				arg0;
 	int				i;
 
 	i = -1;
